@@ -16,14 +16,14 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
 
     defaultConfig {
         applicationId "odbii.starter.automotive.iot.ibm.com.iot4a_odbii"
         minSdkVersion 21
-        targetSdkVersion 28
-        versionCode 8
-        versionName "2.0.0"
+        targetSdkVersion 29
+        versionCode 9
+        versionName "2.0.1"
         multiDexEnabled true
     }
     buildTypes {
@@ -37,9 +37,8 @@ android {
     }
 
     applicationVariants.all{ variant ->
-        def buildDate = new Date().format("yyyyMMdd")
         def buildTimestamp = new Date().format("yyyyMMdd_HHmmss")
-        def versionLabel = "Version: " + variant.versionName + " Build: " + buildDate
+        def versionLabel = "Version: " + variant.versionName + " Build: " + buildTimestamp
         variant.resValue "string", "versionLabel", versionLabel
         variant.outputs.all { output ->
             if(output.outputFileName != null && output.outputFileName.endsWith('.apk')){
@@ -70,11 +69,11 @@ dependencies {
     implementation('org.eclipse.paho:org.eclipse.paho.client.mqttv3:1.2.2') {
         exclude module: 'support-v4'
     }
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support:mediarouter-v7:28.0.0'
-    implementation 'com.android.support:palette-v7:28.0.0'
-    implementation 'com.android.support:support-v4:28.0.0'
-    implementation 'com.android.support:design:28.0.0'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'androidx.mediarouter:mediarouter:1.0.0'
+    implementation 'androidx.palette:palette:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
     implementation 'com.github.pires:obd-java-api:1.0'
     implementation 'com.google.android.gms:play-services:12.0.1'
     implementation 'com.ibm.messaging:watson-iot:0.2.6'
