@@ -19,8 +19,8 @@ package obdii.starter.automotive.iot.ibm.com.iot4a_obdii.settings;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.MenuItem;
 
 public class AppSettingsActivity extends AppCompatActivity {
